Improve Telescope configuration
This is based largely on the configuration recipe linked below and
changes Telescope to search within and find hidden files. Files ignored
by version control should still be ignored.

This also removes the ".git" entry from Telescope's
file_ignore_patterns since that was ignoring any file that include "git"
which was not the intention.

https://github.com/nvim-telescope/telescope.nvim/wiki/Configuration-Recipes#file-and-text-search-in-hidden-files-and-directories